requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
sqlalchemy-utils
psycopg2-binary
python-dotenv
pydantic
python-dotenv